move common steps out of folder
