+++ src/pages/integration-builder/builder/commonSteps/getUserInfo.mdx
