+++ src/pages/integration-builder/builder/commonSteps/installation/installationSolana.mdx
