+++ src/pages/integration-builder/builder/commonSteps/installation/installationStarkEx.mdx
