+++ src/pages/integration-builder/builder/commonSteps/instantiateSDKWhitelabeled.mdx
