+++ src/pages/integration-builder/builder/commonSteps/registerApp.mdx
