+++ src/pages/integration-builder/builder/frameworks/vue/index.ts
@@ -9,31 +9,31 @@ import {
   getRPCFunctionsReturnsVue,
   PLACEHOLDERS,
 } from "../../commonSnippets";
+import * as customAuthenticationStep from "../../commonSteps/customAuthenticationStep.mdx";
+import * as getUserInfo from "../../commonSteps/getUserInfo.mdx";
+import * as importModules from "../../commonSteps/importModules.mdx";
+import * as importModulesCustom from "../../commonSteps/importModulesCustom.mdx";
+import * as initialize from "../../commonSteps/initialize.mdx";
+import * as installation from "../../commonSteps/installation/installation.mdx";
+import * as installationCustom from "../../commonSteps/installation/installationCustom.mdx";
+import * as installationEthers from "../../commonSteps/installation/installationEthers.mdx";
+import * as installationSolana from "../../commonSteps/installation/installationSolana.mdx";
+import * as installationStarkEx from "../../commonSteps/installation/installationStarkEx.mdx";
+import * as installationStarkNet from "../../commonSteps/installation/installationStarkNet.mdx";
+import * as installationTezos from "../../commonSteps/installation/installationTezos.mdx";
+import * as installationWeb3 from "../../commonSteps/installation/installationWeb3.mdx";
+import * as instantiateSDK from "../../commonSteps/instantiateSDK.mdx";
+import * as instantiateSDKWhitelabeled from "../../commonSteps/instantiateSDKWhitelabeled.mdx";
+import * as login from "../../commonSteps/login.mdx";
+import * as logout from "../../commonSteps/logout.mdx";
+import * as registerApp from "../../commonSteps/registerApp.mdx";
+import * as evmRPCFunctions from "../../commonSteps/rpcFunctions/evmRPCFunctions.mdx";
+import * as solanaRPCFunctions from "../../commonSteps/rpcFunctions/solanaRPCFunctions.mdx";
+import * as starkExRPCFunctions from "../../commonSteps/rpcFunctions/starkExRPCFunctions.mdx";
+import * as starkNetRPCFunctions from "../../commonSteps/rpcFunctions/starkNetRPCFunctions.mdx";
+import * as tezosRPCFunctions from "../../commonSteps/rpcFunctions/tezosRPCFunctions.mdx";
+import * as whiteLabeling from "../../commonSteps/whitelabeling.mdx";
 import { ReplaceFileAggregator, toSteps } from "../../utils";
-import * as customAuthenticationStep from "../common/customAuthenticationStep.mdx";
-import * as getUserInfo from "../common/getUserInfo.mdx";
-import * as importModules from "../common/importModules.mdx";
-import * as importModulesCustom from "../common/importModulesCustom.mdx";
-import * as initialize from "../common/initialize.mdx";
-import * as installation from "../common/installation/installation.mdx";
-import * as installationCustom from "../common/installation/installationCustom.mdx";
-import * as installationEthers from "../common/installation/installationEthers.mdx";
-import * as installationSolana from "../common/installation/installationSolana.mdx";
-import * as installationStarkEx from "../common/installation/installationStarkEx.mdx";
-import * as installationStarkNet from "../common/installation/installationStarkNet.mdx";
-import * as installationTezos from "../common/installation/installationTezos.mdx";
-import * as installationWeb3 from "../common/installation/installationWeb3.mdx";
-import * as instantiateSDK from "../common/instantiateSDK.mdx";
-import * as instantiateSDKWhitelabeled from "../common/instantiateSDKWhitelabeled.mdx";
-import * as login from "../common/login.mdx";
-import * as logout from "../common/logout.mdx";
-import * as registerApp from "../common/registerApp.mdx";
-import * as evmRPCFunctions from "../common/rpcFunctions/evmRPCFunctions.mdx";
-import * as solanaRPCFunctions from "../common/rpcFunctions/solanaRPCFunctions.mdx";
-import * as starkExRPCFunctions from "../common/rpcFunctions/starkExRPCFunctions.mdx";
-import * as starkNetRPCFunctions from "../common/rpcFunctions/starkNetRPCFunctions.mdx";
-import * as tezosRPCFunctions from "../common/rpcFunctions/tezosRPCFunctions.mdx";
-import * as whiteLabeling from "../common/whitelabeling.mdx";
 import * as buildingApp from "./buildingApp.mdx";
 import * as webpackIssues from "./webpackIssues.mdx";
